Remove some dependencies inside rocketchat-lib/lib (#13212)
* Move some functions from lib to rc-ui-utils

* Move some functions from lib ro rc-utils

* Remove hostname from RC namespace

* Import variables to remove dependency of RC namespace
