+++ packages/rocketchat-grant/server/grant.js
@@ -1,5 +1,4 @@
-import { RocketChat } from 'meteor/rocketchat:lib';
-
+import { hostname } from 'meteor/rocketchat:lib';
 import Providers from './providers';
 import Settings from './settings';
 import { path, generateCallback, generateAppCallback } from './routes';
@@ -38,7 +37,7 @@ const config = {};
 export function generateConfig() {
 	config.server = {
 		protocol: 'http',
-		host: RocketChat.hostname,
+		host: hostname,
 		path,
 		state: true,
 	};
