+++ packages/rocketchat-lib/client/lib/roomExit.js
@@ -1,7 +1,7 @@
 import { Blaze } from 'meteor/blaze';
 import { BlazeLayout } from 'meteor/kadira:blaze-layout';
 import { callbacks } from 'meteor/rocketchat:callbacks';
-import { currentTracker } from './openRoom';
+import { currentTracker } from 'meteor/rocketchat:ui-utils';
 
 this.roomExit = function() {
 	// 7370 - Close flex-tab when opening a room on mobile UI
