+++ packages/rocketchat-lib/lib/MessageTypes.js
@@ -1,10 +1,11 @@
 import { Meteor } from 'meteor/meteor';
 import { MessageTypes } from 'meteor/rocketchat:ui-utils';
+import { callbacks } from 'meteor/rocketchat:callbacks';
 
 RocketChat.MessageTypes = MessageTypes;
 
 Meteor.startup(function() {
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'r',
 		system: true,
 		message: 'Room_name_changed',
@@ -15,7 +16,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'au',
 		system: true,
 		message: 'User_added_by',
@@ -26,7 +27,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'ru',
 		system: true,
 		message: 'User_removed_by',
@@ -37,7 +38,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'ul',
 		system: true,
 		message: 'User_left',
@@ -47,7 +48,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'uj',
 		system: true,
 		message: 'User_joined_channel',
@@ -57,7 +58,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'wm',
 		system: true,
 		message: 'Welcome',
@@ -67,7 +68,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'rm',
 		system: true,
 		message: 'Message_removed',
@@ -77,13 +78,13 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'rtc',
 		render(message) {
-			return RocketChat.callbacks.run('renderRtcMessage', message);
+			return callbacks.run('renderRtcMessage', message);
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'user-muted',
 		system: true,
 		message: 'User_muted_by',
@@ -94,7 +95,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'user-unmuted',
 		system: true,
 		message: 'User_unmuted_by',
@@ -105,7 +106,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'subscription-role-added',
 		system: true,
 		message: '__username__was_set__role__by__user_by_',
@@ -117,7 +118,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'subscription-role-removed',
 		system: true,
 		message: '__username__is_no_longer__role__defined_by__user_by_',
@@ -129,7 +130,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'room-archived',
 		system: true,
 		message: 'This_room_has_been_archived_by__username_',
@@ -139,7 +140,7 @@ Meteor.startup(function() {
 			};
 		},
 	});
-	RocketChat.MessageTypes.registerType({
+	MessageTypes.registerType({
 		id: 'room-unarchived',
 		system: true,
 		message: 'This_room_has_been_unarchived_by__username_',
