+++ packages/rocketchat-lib/lib/roomTypes/favorite.js
@@ -1,4 +1,6 @@
 import { Meteor } from 'meteor/meteor';
+import { settings } from 'meteor/rocketchat:settings';
+import { getUserPreference } from 'meteor/rocketchat:utils';
 import { RoomTypeConfig } from '../RoomTypeConfig';
 
 export class FavoriteRoomType extends RoomTypeConfig {
@@ -12,6 +14,6 @@ export class FavoriteRoomType extends RoomTypeConfig {
 		});
 	}
 	condition() {
-		return RocketChat.settings.get('Favorite_Rooms') && RocketChat.getUserPreference(Meteor.userId(), 'sidebarShowFavorites');
+		return settings.get('Favorite_Rooms') && getUserPreference(Meteor.userId(), 'sidebarShowFavorites');
 	}
 }
