+++ packages/rocketchat-lib/lib/roomTypes/private.js
@@ -1,6 +1,10 @@
 import { Meteor } from 'meteor/meteor';
-import { RoomSettingsEnum, RoomTypeConfig, RoomTypeRouteConfig, UiTextContext } from '../RoomTypeConfig';
 import { ChatRoom } from 'meteor/rocketchat:models';
+import { openRoom } from 'meteor/rocketchat:ui-utils';
+import { settings } from 'meteor/rocketchat:settings';
+import { hasAtLeastOnePermission, hasPermission } from 'meteor/rocketchat:authorization';
+import { getUserPreference } from 'meteor/rocketchat:utils';
+import { RoomSettingsEnum, RoomTypeConfig, RoomTypeRouteConfig, UiTextContext } from '../RoomTypeConfig';
 
 export class PrivateRoomRoute extends RoomTypeRouteConfig {
 	constructor() {
@@ -36,7 +40,7 @@ export class PrivateRoomType extends RoomTypeConfig {
 	}
 
 	roomName(roomData) {
-		if (RocketChat.settings.get('UI_Allow_room_names_with_special_chars')) {
+		if (settings.get('UI_Allow_room_names_with_special_chars')) {
 			return roomData.fname || roomData.name;
 		}
 
@@ -44,8 +48,8 @@ export class PrivateRoomType extends RoomTypeConfig {
 	}
 
 	condition() {
-		const groupByType = RocketChat.getUserPreference(Meteor.userId(), 'sidebarGroupByType');
-		return groupByType && RocketChat.authz.hasPermission('view-p-room');
+		const groupByType = getUserPreference(Meteor.userId(), 'sidebarGroupByType');
+		return groupByType && hasPermission('view-p-room');
 	}
 
 	isGroupChat() {
@@ -53,7 +57,7 @@ export class PrivateRoomType extends RoomTypeConfig {
 	}
 
 	canAddUser(room) {
-		return RocketChat.authz.hasAtLeastOnePermission(['add-user-to-any-p-room', 'add-user-to-joined-room'], room._id);
+		return hasAtLeastOnePermission(['add-user-to-any-p-room', 'add-user-to-joined-room'], room._id);
 	}
 
 	allowRoomSettingChange(room, setting) {
@@ -68,7 +72,7 @@ export class PrivateRoomType extends RoomTypeConfig {
 				return !room.broadcast && room.ro;
 			case RoomSettingsEnum.SYSTEM_MESSAGES:
 			case RoomSettingsEnum.E2E:
-				return RocketChat.settings.get('E2E_Enable') === true;
+				return settings.get('E2E_Enable') === true;
 			default:
 				return true;
 		}
