+++ packages/rocketchat-ui-utils/client/index.js
@@ -23,3 +23,5 @@ export { popout } from './lib/popout';
 export { messageProperties } from '../lib/MessageProperties';
 export { MessageTypes } from '../lib/MessageTypes';
 export { alerts } from './lib/alerts';
+export { Message } from '../lib/Message';
+export { currentTracker, openRoom } from './lib/openRoom';
