+++ packages/rocketchat-ui-utils/server/index.js
@@ -1,2 +1,3 @@
 export { messageProperties } from '../lib/MessageProperties';
 export { MessageTypes } from '../lib/MessageTypes';
+export { Message } from '../lib/Message';
